addCustomer method implemented but not tested. Login.java and Registration.java made into child classes of UserManagement.java for abstraction.
