+++ src/main/java/com/CSI2132Project/Registration.java
@@ -1,0 +1,14 @@
+package com.CSI2132Project;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.ResultSetMetaData;
+
+public class Registration extends UserManagement{
+     /*
+    The functionality for this class will be similar to login, in that it will check if a user is already registered. If they are, they will not be allowed
+    to register again. If they are not, they will be allowed to register and it will setup a new user in the database.
+     */
+
+}
